Add support for numpy numbers to be json serializable
The problem was discovered by running
https://github.com/combust/mleap-demo/blob/master/notebooks/airbnb-price-regression-scikit.ipynb

During a bundle serialization an object was constructed
to describe a model. This object for some reason contained numpy numbers
instead of python native integers.

Therefore it was not json serializable cause the type was not recognized

This patch adds support for numpy numbers for json serializer